+++ lib/src/Test/Falsify/GenDefault/Std.hs
@@ -1,13 +1,16 @@
+-- | \"Standard\" default generators
 module Test.Falsify.GenDefault.Std
   ( Std
   ) where
 
-import Test.Falsify.GenDefault (ViaIntegral (..), GenDefault, ViaEnum (..), ViaGeneric (..))
-import Data.Int (Int8, Int16, Int32, Int64)
-import Data.Word (Word8, Word16, Word32, Word64)
+import Data.Int
+import Data.Word
+
+import Test.Falsify.GenDefault
 
 -- | Type tag for these "standard" default generators.
--- You can use this tag directly or choose type-by-type with 'ViaTag'.
+--
+-- You can use this tag directly or choose type-by-type with t'ViaTag'.
 data Std
 
 deriving via (ViaEnum ()) instance GenDefault Std ()